working on notification flow/layout
