+++ src/consts/walletConnect.ts
@@ -1,0 +1,7 @@
+export const WC_NOTIFICATION_TYPES = [
+  'connect',
+  'disconnect',
+  'session_request',
+  'provenance_sign',
+  'provenance_sendTransaction',
+];
